Rename 'Driver' to 'UpperLayer'
'Driver' trait naming choice felt confusing and did not really represents its intents.

- Rename Driver trait to 'UpperLayer'
- Move trait outside of "phy" module
- Renamed methods : 'transmit' -> 'frame_to_transmit', 'received' -> 'received_frame'
- Remove 'driver' source file.
- Updated tests for Upperlayer
