+++ dot15d4/src/lib.rs
@@ -13,3 +13,4 @@ pub mod mac;
 pub mod phy;
 pub mod sync;
 pub mod time;
+pub mod upper;
